[Test] Split out the @objc bits of the new generic_casts.swift code.
rdar://problem/56044443
